Add support for string literals
Allow string literals to be converted to enums  Example:
  const type Direction = "NORTH" | "EAST" | "SOUTH" | "WEST";
    to become
  enum Direction {
    NORTH
    EAST
    SOUTH
    WEST
  }

This gives us the ability to have stronger types than enums support and
avoids us having to cast between enums when we want to move from a specific
type to a general type where the specific type is a subset of the general
type.  Example:
const enum Colors {
  BLACK = 'BLACK',
  RED = 'RED',
  ORANGE = 'ORANGE',
  YELLOW = 'YELLOW',
  GREEN = 'GREEN',
  BLUE = 'BLUE',
  INDIGO = 'INDIGO',
  VIOLET = 'VIOLET',
  WHITE = 'WHITE',
}

const PrimaryColors {
  RED = 'RED',
  GREEN = 'GREEN',
  BLUE = 'BLUE'
}

Requires doing:
  const inputColor:PrimaryColors = PrimaryColors.RED;
  const myColor:Colors = Colors[inputColor.toString()];

Now you can define PrimaryColors as a string literal union and Typescript
will know the typing.  And we can map them to GraphQL